+++ router.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * COPS (Calibre OPDS PHP Server) router script for the PHP built-in web server @todo
+ * COPS (Calibre OPDS PHP Server) router script for the PHP built-in web server
  * with $config['cops_use_route_urls'] = '1' and no PHP script in URL (dev only)
  *
  * $ php -S 0.0.0.0:8080 router.php
@@ -15,19 +15,20 @@ if (php_sapi_name() !== 'cli-server') {
     return;
 }
 
-// @todo route to the the right endpoint if needed
-//$path = urldecode($_SERVER['SCRIPT_NAME']);
+// check if the requested path actually exists
 $path = parse_url((string) $_SERVER['REQUEST_URI'], PHP_URL_PATH);
 if (!empty($path) && file_exists(__DIR__ . $path) && !is_dir(__DIR__ . $path)) {
     return false;
 }
+// route to the right PHP endpoint if needed
 $script = urldecode((string) $_SERVER['SCRIPT_NAME']);
 if (str_contains($path, $script . '/') && file_exists(__DIR__ . $script)) {
     return false;
 }
 
+// set environment vars for the front controller
 $_SERVER['SCRIPT_NAME'] = '/index.php';
 $_SERVER['PATH_INFO'] ??= parse_url((string) $_SERVER['REQUEST_URI'], PHP_URL_PATH);
 
-// @todo do not use index.php here yet
+// use index.php as front controller
 include __DIR__ . '/index.php';