+++ src/Handlers/CheckHandler.php
@@ -21,6 +21,7 @@ class CheckHandler extends BaseHandler
     public static function getRoutes()
     {
         return [
+            "/check/{more:.*}" => [static::PARAM => static::HANDLER],
             "/check" => [static::PARAM => static::HANDLER],
         ];
     }